Move general functions to utils.py
